# Add basic benchmark