--- a/benchmarks/CMakeLists.txt
+++ b/benchmarks/CMakeLists.txt
@@ -1,0 +1,27 @@
+project(chatterino-benchmark)
+
+set(benchmark_SOURCES
+    ${CMAKE_CURRENT_LIST_DIR}/src/main.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/src/Emojis.cpp
+    )
+
+add_executable(${PROJECT_NAME} ${benchmark_SOURCES})
+add_sanitizers(${PROJECT_NAME})
+
+target_link_libraries(${PROJECT_NAME} PRIVATE chatterino-lib)
+
+target_link_libraries(${PROJECT_NAME} PRIVATE benchmark::benchmark)
+
+target_compile_definitions(${PROJECT_NAME} PRIVATE
+    CHATTERINO_TEST
+    )
+
+set_target_properties(${PROJECT_NAME}
+    PROPERTIES
+    ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib"
+    LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib"
+    RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin"
+    RUNTIME_OUTPUT_DIRECTORY_RELEASE "${CMAKE_BINARY_DIR}/bin"
+    RUNTIME_OUTPUT_DIRECTORY_DEBUG "${CMAKE_BINARY_DIR}/bin"
+    RUNTIME_OUTPUT_DIRECTORY_RELWITHDEBINFO "${CMAKE_BINARY_DIR}/bin"
+    )