Pull request checklist:

- [x] `CHANGELOG.md` was updated, if applicable

# Description

Same benchmark as #2816, except 'modernized' with the cmake changes made some time ago

<!-- If applicable, please include a summary of what you've changed and what issue is fixed. In the case of a bug fix, please include steps to reproduce the bug so the pull request can be tested -->