@pajlada 2021-07-18
@Felanbird Could you proofread my english & sentence structure here? https://github.com/Chatterino/chatterino2/blob/chore/benchmarks/docs/test-and-benchmark.md#test-and-benchmark

Thank you!