@zneix 2021-08-08
Before updating the branch
```
zneix@uds ~/g/c/build (chore/benchmarks)$ ./bin/chatterino-benchmark
2021-08-08T13:47:13+02:00
Running ./bin/chatterino-benchmark
Run on (16 X 2900 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 4096 KiB (x2)
Load Average: 8.44, 8.26, 5.10
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
BM_ShortcodeParsing       7583 ns         7433 ns        89671
```

After updating the branch
```zneix@uds ~/g/c/build (chore/benchmarks)$ ./bin/chatterino-benchmark
2021-08-08T13:55:53+02:00
Running ./bin/chatterino-benchmark
Run on (16 X 2900 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 4096 KiB (x2)
Load Average: 3.65, 7.07, 5.78
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
BM_ShortcodeParsing       7500 ns         7482 ns        86425
```